+++ packages/generator-converters/src/wireQueryDataTypeToQueryDataTypeDefinition.ts
@@ -161,6 +161,8 @@ export function wireQueryDataTypeToQueryDataTypeDefinition<
 
     case "null":
     case "unsupported":
+    case "mediaReference":
+    case "void":
       throw new Error(
         `Unable to process query because the server indicated an unsupported QueryDataType.type: ${input.type}. Please check that your query is using supported types.`,
       );