+++ packages/react-components/src/object-table/ObjectTable.tsx
@@ -60,8 +60,10 @@ export function ObjectTable<
   >,
 >({
   objectType,
+  objectSet,
   columnDefinitions,
   filter,
+  objectSetOptions,
   orderBy,
   defaultOrderBy,
   onOrderByChanged,
@@ -119,6 +121,8 @@ export function ObjectTable<
     columnDefinitions,
     filter,
     sorting,
+    objectSet,
+    objectSetOptions,
   );
 
   const { columns, loading: isColumnsLoading } = useColumnDefs<