+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.112
+
+### Patch Changes
+
+- Updated dependencies [4470c0d]
+  - @osdk/client@2.8.0-beta.13
+
+## 0.1.0-beta.111
+
+### Patch Changes
+
+- Updated dependencies [8825f8c]
+  - @osdk/client@2.8.0-beta.12
+
 ## 0.1.0-beta.110
 
 ### Patch Changes