+++ packages/create-app.template.tutorial-todo-aip-app.beta/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-aip-app.beta
 
+## 2.8.0-beta.13
+
+## 2.8.0-beta.12
+
 ## 2.8.0-beta.11
 
 ## 2.8.0-beta.10