+++ packages/generator-converters.ontologyir/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/generator-converters.ontologyir
 
+## 2.8.0-beta.13
+
+### Patch Changes
+
+- @osdk/client.unstable@2.8.0-beta.13
+
+## 2.8.0-beta.12
+
+### Patch Changes
+
+- @osdk/client.unstable@2.8.0-beta.12
+
 ## 2.8.0-beta.11
 
 ### Patch Changes