+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osdk/generator-converters
 
+## 2.8.0-beta.13
+
+### Minor Changes
+
+- 4470c0d: Add support to execute queries on branches.
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.13
+
+## 2.8.0-beta.12
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.12
+
 ## 2.8.0-beta.11
 
 ### Patch Changes